COG-628:  checkCanEditRoles(UserType cognitoUser) and checkCanEditPermissions(UserType cognitoUser) methods are removed from AuthorizationService
