+++ src/main/java/gov/ca/cwds/idm/service/authorization/AuthorizationService.java
@@ -2,6 +2,7 @@ package gov.ca.cwds.idm.service.authorization;
 
 import com.amazonaws.services.cognitoidp.model.UserType;
 import gov.ca.cwds.idm.dto.User;
+import gov.ca.cwds.idm.dto.UserUpdate;
 
 public interface AuthorizationService {
 
@@ -9,7 +10,7 @@ public interface AuthorizationService {
 
   void checkCanCreateUser(User user);
 
-  void checkCanUpdateUser(UserType existingUser);
+  void checkCanUpdateUser(UserType existingUser, UserUpdate updateUserDto);
 
   void checkCanUpdateUser(String userId);
 
@@ -17,11 +18,7 @@ public interface AuthorizationService {
 
   void checkCanResendInvitationMessage(String id);
 
-  void checkCanEditRoles(UserType cognitoUser);
-
   boolean canEditRoles(User user);
 
-  void checkCanEditPermissions(UserType cognitoUser);
-
-  boolean canEditPermissions(User  user);
+  boolean canEditPermissions(User user);
 }
